Ensure incoming sensorScanPosition is integer consistently (#2916)
* ensure incoming sensorScanPosition is integer, converted to float/real for CRMT only

* fix my bug for header file float turned int

* Empty-Commit

* change to use iFov as integer per Ben Johnson

* fix scan angle code to be flexible for either float or int since sometimes it is sensorViewAngle (a float) and sometimes sensorScanPosition (an int)

* kick off tests with new ufo-data

* kick off test again after ufo-data merge

* empty commit kick off regression tests

* Empty-Commit

---------

+++ src/ufo/predictors/CloudLiquidWater.h
@@ -89,7 +89,7 @@ class CloudLiquidWater : public PredictorBase {
                                       const std::vector<float> &,
                                       const std::vector<float> &,
                                       const std::vector<float> &,
-                                      const std::vector<float> &,
+                                      const std::vector<int> &,
                                       const std::vector<std::vector<std::vector<float>>> &,
                                       const std::vector<std::vector<float>> &,
                                       const std::vector<float> &,
